requirements.txt
pytest==7.4.3
allure-pytest==2.13.2
requests==2.32.4
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
faker==20.1.0
pydantic==2.11.7
playwright==1.54.0
selenium==4.34.2
kafka-python==2.2.15
grpcio==1.74.0
grpcio-tools==1.74.0
protobuf==6.31.1
zeep==4.3.1
lxml==6.0.0
stomp.py==8.2.0
gql==3.4.1
requests-toolbelt==1.0.0
pytest-xdist==3.5.0
PyPDF2==3.0.1
pandas==2.2.3
openpyxl==3.1.5
pytest-metadata==3.1.1
pytest-json-report==1.5.0
pytest-rerunfailures==16.0.1